Release 0.9.2 full arome 3dvar components and reproductibility

+++ src/vortex/tools/odb.py
@@ -13,85 +13,164 @@ from vortex import tools
 from . import addons
 
 
-def setup_env(env=None, date=None, npool=1, nslots=1, iomethod=1, cma='ECMA'):
-    """Setup given environment with default ODB env variables."""
+class OdbDriver(object):
+    """A dedicated class for handling some ODB settings."""
 
-    env.update(
-        ODB_CMA                = cma,
-        BATOR_NBPOOL           = npool,
-        BATODB_NBPOOL          = npool,
-        BATOR_NBSLOT           = nslots,
-        BATODB_NBSLOT          = nslots,
-        ODB_IO_METHOD          = iomethod,
-        ODB_DEBUG              = env.ODB_DEBUG              or 0,
-        ODB_CTX_DEBUG          = env.ODB_CTX_DEBUG          or 0,
-        ODB_REPRODUCIBLE_SEQNO = env.ODB_REPRODUCIBLE_SEQNO or 4,
-        ODB_STATIC_LINKING     = env.ODB_STATIC_LINKING     or 1,
-        ODB_ANALYSIS_DATE      = date.ymd,
-        ODB_ANALYSIS_TIME      = env.ODB_ANALYSIS_TIME      or date.hm + '00',
-        TO_ODB_ECMWF           = 0,
-        TO_ODB_SWAPOUT         = env.TO_ODB_SWAPOUT         or 0,
-    )
+    def __init__(self, sh=None, env=None, target=None):
+        """A quite challenging initialisation since sh, env and target should be provided."""
+        self.sh     = sh
+        if self.sh is None:
+            logger.critical('%s created with a proper shell access [%s]', self.__class__, self)
+        self.env    = env
+        if self.env is None:
+            logger.critical('%s created with a proper environment access [%s]', self.__class__, self)
+        self.target = target
+        if self.target is None:
+            logger.critical('%s created with a proper target access [%s]', self.__class__, self)
 
-    if iomethod == 4:
-        env.update(
-            ODB_IO_GRPSIZE  = env.ODB_IO_GRPSIZE  or npool,
-            ODB_IO_FILESIZE = env.ODB_IO_FILESIZE or 128,
+    def setup(self, date=None, npool=1, nslot=1, iomethod=1, layout='ecma'):
+        """Setup given environment with default ODB env variables."""
+
+        self.env.update(
+            ODB_CMA                = layout.upper(),
+            BATOR_NBPOOL           = npool,
+            BATODB_NBPOOL          = npool,
+            BATOR_NBSLOT           = nslot,
+            BATODB_NBSLOT          = nslot,
+            ODB_IO_METHOD          = iomethod,
         )
+
+        self.env.default(
+            ODB_DEBUG              = 0,
+            ODB_CTX_DEBUG          = 0,
+            ODB_REPRODUCIBLE_SEQNO = 4,
+            ODB_STATIC_LINKING     = 1,
+            ODB_ANALYSIS_DATE      = date.ymd,
+            ODB_ANALYSIS_TIME      = date.hm + '00',
+            TO_ODB_ECMWF           = 0,
+            TO_ODB_SWAPOUT         = 0,
+        )
+
+        if iomethod == 4:
+            self.env.default(
+                ODB_IO_GRPSIZE  = npool,
+                ODB_IO_FILESIZE = 128,
+        )
+
+        if self.sh.path.exists('IOASSIGN'):
+            self.env.default(
+                IOASSIGN = self.sh.path.abspath('IOASSIGN'),
+            )
+
+    def ioassign_create(self, ioassign='ioassign.x', npool=1, layout='ecma'):
+        """Build IO-Assign table."""
+        iopath   = self.target.get('odbtools:rootdir', self.env.TMPDIR)
+        iovers   = self.target.get('odbtools:odbcycle', 'oper')
+        iocreate = self.target.get('odbtools:iocreate', 'create_ioassign')
+        iocmd    = self.env.get('ODB_IOCREATE_COMMAND', self.sh.path.join(iopath, iovers, iocreate))
+        ioassign = self.sh.path.abspath(ioassign)
+        self.sh.chmod(ioassign, 0755)
+        self.env.ODB_IOASSIGN_BINARY = ioassign
+        self.sh.spawn([iocmd, '-l' + layout.upper(), '-n' + str(npool)], output=False)
+
+    def ioassign_merge(self, ioassign='ioassign.x', layout='ecma', odbnames=None):
+        """Build IO-Assign table."""
+        layout   = layout.upper()
+        iopath   = self.target.get('odbtools:rootdir', self.env.TMPDIR)
+        iovers   = self.target.get('odbtools:odbcycle', 'oper')
+        iocreate = self.target.get('odbtools:iocreate', 'create_ioassign')
+        iomerge  = self.target.get('odbtools:iomerge', 'create_ioassign')
+        iocmd    = [self.env.get('ODB_IOMERGE_COMMAND', self.sh.path.join(iopath, iovers, iomerge))]
+        self.sh.chmod(ioassign, 0755)
+        self.env.default(
+            ODB_IOASSIGN_BINARY  = self.sh.path.abspath(ioassign),
+            ODB_IOCREATE_COMMAND = self.sh.path.join(iopath, iovers, iocreate),
+        )
+        oldpwd = self.sh.getcwd()
+        self.sh.cd(layout, create=True)
+        self.env.ODB_SRCPATH_ECMA = self.env.ODB_DATAPATH_ECMA = self.sh.getcwd()
+        iocmd.extend(['-d', oldpwd])
+        for dbname in odbnames:
+            iocmd.extend(['-t', dbname])
+        self.sh.spawn(iocmd, output=False)
+        self.sh.cd(oldpwd)
+
+
+class OdbComponent(object):
+    """Extend Algo Components with ODB features."""
+
+    @property
+    def odb(self):
+        if not hasattr(self, '_odb'):
+            self._odb = OdbDriver(
+                sh     = self.system,
+                env    = self.env,
+                target = self.target,
+            )
+        return self._odb
 
 
 class TimeSlots(object):
-    def __init__(self, nslots=7, start='-PT3H', window='PT6H', chunk='PT1H', center=True):
-        if isinstance(nslots, str):
-            info = [ x.strip() for x in nslots.split('/') ]
-            nslots = info[0]
+    def __init__(self, nslot=7, start='-PT3H', window='PT6H', chunk='PT1H', center=True):
+        if isinstance(nslot, str):
+            info = [ x.strip() for x in nslot.split('/') ]
+            nslot = info[0]
             if len(info) > 1:
                 start = info[1]
             if len(info) > 2:
                 window = info[2]
             if len(info) > 3:
                 chunk = info[3]
-        self.nslots = int(nslots)
+        self.nslot  = int(nslot)
         self.start  = tools.date.Period(start)
         self.window = tools.date.Period(window)
-        self.chunk  = self.window if self.nslots == 1 else tools.date.Period(chunk)
+        self.chunk  = self.window if self.nslot < 2 else tools.date.Period(chunk)
         self.center = center
 
-    def at_date(self, date):
-        """Return a list of date bounds for the slots at this ``date``."""
+    def as_slots(self):
+        """Return a list of slots in seconds."""
         if self.center:
-            self.slots = [ self.chunk.length for x in range(self.nslots) ]
+            slots = [ self.chunk.length for x in range(self.nslot) ]
             nb = int(self.window.length / self.chunk.length)
-            if nb != self.nslots:
-                self.slots[0] = self.slots[-1] = self.chunk.length / 2
+            if nb != self.nslot:
+                slots[0] = slots[-1] = self.chunk.length / 2
         else:
-            islot = int(self.window.length / self.nslots)
-            self.slots = [ islot for x in range(self.nslots) ]
+            islot = int(self.window.length / self.nslot)
+            slots = [ islot for x in range(self.nslot) ]
+        return slots
 
-    def as_list(self, date, filename=None):
+    def as_bounds(self, date, filename=None):
         """Return time slots as a list of compact date values."""
-
         date = tools.date.Date(date)
-        self.at_date(date)
-
         boundlist = [ date + self.start ]
-        for x in self.slots:
+        for x in self.as_slots():
             boundlist.append(boundlist[-1] + x)
         boundlist = [ x.compact() for x in boundlist ]
 
         return boundlist
 
+    def leftmargin(self, date):
+        """Return length in minutes from left margin of the window."""
+        return int(self.start.total_seconds() / 60)
+
+    def rightmargin(self, date):
+        """Return length in minutes from rigth margin of the window."""
+        date = tools.date.Date(date)
+        end = date + self.start + self.window
+        delta = end - date
+        return int(delta.total_seconds() / 60)
+
     def as_file(self, date, filename):
         """Fill the specified ``filename`` wih the current list of time slots at this ``date``."""
         nbx = 0
         with io.open(filename, 'w') as fd:
-            for x in self.as_list(date):
+            for x in self.as_bounds(date):
                 fd.write(unicode(x + '\n'))
             nbx = fd.tell()
         return nbx
 
 
-class ODB_Tool(addons.Addon):
+class OdbShell(addons.Addon):
     """
     Default interface to ODB commands.
     These commands extend the shell.
@@ -107,19 +186,27 @@ class ODB_Tool(addons.Addon):
                 optional = True,
                 default  = 'ODB.tgz',
             ),
+            pipeget = dict(
+                type     = bool,
+                optional = True,
+                default  = True,
+            ),
         )
     )
 
     def odb_cp(self, source, destination, intent='in'):
         """Extended copy for ODB repository."""
-        logger.info('Using ODB copy')
         rc, source, destination = self.sh.tarfix_out(source, destination)
         rc = rc and self.sh.cp(source, destination, intent=intent)
-        rt, source, destination = self.sh.tarfix_in(source, destination)
-        if intent == 'inout':
-            for infile in self.ffind(destination):
-                self.chmod(linkedfile, 0644)
-        return rc and rt
+        if rc :
+            rc, source, destination = self.sh.tarfix_in(source, destination)
+            if rc and intent == 'inout':
+                self.sh.stderr('chmod', 0644, destination)
+                oldtrace, self.sh.trace = self.sh.trace, False
+                for infile in self.sh.ffind(destination):
+                    self.sh.chmod(infile, 0644)
+                self.sh.trace = oldtrace
+        return rc
 
     def odb_credentials(self, hostname=None, logname=None):
         """Some heuristic to get proper values for these arguments."""
@@ -150,10 +237,20 @@ class ODB_Tool(addons.Addon):
             loccwd = self.sh.getcwd()
             loctmp = tempfile.mkdtemp(prefix='odb_', dir=loccwd)
             self.sh.cd(loctmp)
-            rc = ftp.get(source, self.tmpname)
+            if self.pipeget:
+                p = self.sh.popen(
+                    ['tar', 'xvfz', '-'],
+                    stdin   = True,
+                    output  = False,
+                    bufsize = 8192,
+                )
+                rc = ftp.get(source, p.stdin)
+                self.sh.pclose(p)
+            else:
+                rc = ftp.get(source, self.tmpname)
+                self.sh.untar(self.tmpname)
+                self.sh.rm(self.tmpname)
             ftp.close()
-            self.sh.untar(self.tmpname)
-            self.sh.rm(self.tmpname)
             unpacked = self.sh.glob('*')
             self.sh.mv(unpacked[-1], destination)
             self.sh.cd(loccwd)
@@ -176,6 +273,7 @@ class ODB_Tool(addons.Addon):
         if ftp:
             p = self.sh.popen(
                 ['tar', 'cvfz', '-', source],
+                stdout  = True,
                 output  = False,
                 bufsize = 8192,
             )
